Remove unused data files and split main script
